+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.8.0] - 2023-11-13
+
+### Added
+
+- Add graceful delay on startup where unhandled errors won't trigger a critical exit 
+
 ## [1.7.0] - 2023-11-07
 
 ### Added